feat(core): extract shared scaffold-hygiene logic, dedup cursor/opencode (closes #97)
ensureGitignore()/ensureFormatterIgnores() (plus private helpers) were
independently hand-duplicated, byte-for-byte except for a parameter-name
difference, across plugins/adlc-cursor and plugins/adlc-opencode. Every
.gitignore edge case found during #92's 5 review rounds had to be fixed in
both copies to keep them in sync.

Extract both functions into packages/core/lib/scaffold-hygiene.mjs,
re-exported from @adlc/core's index (both plugins already depend on and
import from @adlc/core elsewhere). Both plugins' scaffold.mjs now import and
re-export the shared functions instead of reimplementing them.

Consolidate the ~17 duplicated behavioral tests (also byte-for-byte identical
modulo fixture style) into packages/core/test/scaffold-hygiene.test.mjs. Each
plugin's own test file gets one "delegates to the shared implementation" test
asserting reference equality, so a future one-sided edit fails a test instead
of silently drifting.
